+++ src/Test/QuickCheck/Classes/Json.hs
@@ -12,7 +12,7 @@ module Test.QuickCheck.Classes.Json
 
 import Data.Proxy (Proxy)
 import Test.QuickCheck hiding ((.&.))
-import Test.QuickCheck.Property (Property)
+import Test.QuickCheck.Property (Property(..))
 
 #if HAVE_AESON
 import Data.Aeson (FromJSON(..), ToJSON(..))
@@ -46,7 +46,23 @@ jsonEncodingEqualsValue _ = property $ \(a :: a) ->
     Just (v :: AE.Value) -> v == toJSON a
 
 jsonEncodingPartialIsomorphism :: forall a. (ToJSON a, FromJSON a, Show a, Eq a, Arbitrary a) => Proxy a -> Property
-jsonEncodingPartialIsomorphism _ = property $ \(a :: a) ->
-  AE.decode (AE.encode a) == Just a
-
+jsonEncodingPartialIsomorphism _ =
+#if MIN_VERSION_QuickCheck(2,9,0)
+  again $
+#endif
+  MkProperty $
+    arbitrary >>= \(x :: a) ->
+      unProperty $
+      shrinking shrink x $ \x' ->
+        let desc1 = "Just"
+            desc2 = "Data.Aeson.decode . Data.Aeson.encode"
+            name1 = "Data.Aeson.encode a"
+            name2 = "Data.Aeson.decode (Data.Aeson.encode a)"
+            b1  = AE.encode x'
+            b2  = AE.decode (AE.encode x')
+            sb1 = show b1
+            sb2 = show b2
+            description = "  Description: " ++ desc1 ++ " == " ++ desc2
+            err = description ++ "\n" ++ unlines (map ("  " ++) (["a = " ++ show x'])) ++ "  " ++ name1 ++ " = " ++ sb1 ++ "\n  " ++ name2 ++ " = " ++ sb2
+        in counterexample err (Just x' == b2)
 #endif